#242 - Add new 'colortxt' output_format, 'txt' output_format no longer colorizes output
Closes #242
